requirements.txt
livekit-agents>=0.12.0
livekit-plugins-deepgram>=0.8.0
livekit-plugins-openai>=0.10.0
livekit-plugins-elevenlabs>=0.9.0
httpx>=0.27.0
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
matplotlib>=3.8.0
pandas>=2.2.0
click>=8.1.7
tabulate>=0.9.0
simple-salesforce>=1.12.0
hubspot-api-client>=9.0.0
